+++ apps/greenhouse/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @cloudoperators/juno-app-greenhouse
 
+## 0.3.6
+
+### Patch Changes
+
+- a6789db: This change corrects URL state handling for filter property
+- Updated dependencies [a6789db]
+  - @cloudoperators/juno-app-heureka@2.10.1
+
+## 0.3.5
+
+### Patch Changes
+
+- 3810a93: add support_group only if supportGroups exists and is an array
+- Updated dependencies [b1a5eb1]
+- Updated dependencies [c9190d0]
+- Updated dependencies [879812b]
+- Updated dependencies [18ebaba]
+  - @cloudoperators/juno-ui-components@2.30.0
+  - @cloudoperators/juno-app-supernova@0.14.3
+
+## 0.3.4
+
+### Patch Changes
+
+- eea5de6: The pagination is added to the list of component instances for a selected service on the service details page.
+- Updated dependencies [63a951c]
+- Updated dependencies [5a4e3d1]
+- Updated dependencies [7cb142d]
+- Updated dependencies [eea5de6]
+- Updated dependencies [a278544]
+  - @cloudoperators/juno-ui-components@2.29.0
+  - @cloudoperators/juno-app-heureka@2.10.0
+
 ## 0.3.3
 
 ### Patch Changes